handlers/ecs/ngrams/pyproject.toml
[tool.poetry]
name = "ngrams"
version = "0.1.0"
description = ""
authors = ["dev <dev@togglecorp.com>"]

[tool.poetry.dependencies]
python = ">=3.8"
boto3 = "*"
psycopg2-binary = "*"
ngrams_generator = { git = "https://github.com/the-deep-nlp/ngrams-generator.git", rev = "f5d7ecd", branch = "main" }
sentry-sdk = "==1.5.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
